Crate re_blueprint_tree

This crate implements the UI for the blueprint tree in the left panel.

Modules

blueprint_tree 🔒

data

Data structure describing the contents of the blueprint tree

data_result_node_or_path 🔒

Structs

BlueprintTree

Holds the state of the blueprint tree UI.