re_blueprint_tree
Module data

Data structure describing the contents of the blueprint tree

The design goal of these structures is to cover the entire underlying tree (container hierarchy and data result hierarchies) by walking it once entirely, applying filtering on the way.

This is done regardless of whether the data is actually used by the UI (e.g. if everything is collapsed). Benchmarks have indicated that this approach incurs a negligible overhead compared to the overall cost of having large blueprint trees (a.k.a the many-entities performance issues: https://github.com/rerun-io/rerun/issues/8233).

Structs

BlueprintTreeData

Top-level blueprint tree structure.

ContainerData

Data related to a single container and its children.

DataResultData

ViewData

Data related to a single view and its content.

Enums

BlueprintTreeItem

Wrapper structure used for the closure of BlueprintTreeData::visit and friends.

ContentsData

Data for either a container or a view (both of which possible child of a container).

DataResultKind

The various kind of things that may be represented in a data result tree.

Functions

default_open_for_data_result 🔒

If a group or view has a total of this number of elements, show its subtree by default?